+++ scripts/git-subtree/metadata/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/widgets/metadata `
+    https://github.com/Zuoqiu-Yingyi/widget-metadata.git `
+    dev `
+    --squash