+++ scripts/git-subtree/metadata/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/widgets/metadata `
+    https://github.com/Zuoqiu-Yingyi/widget-metadata.git `
+    dev