+++ scripts/git-subtree/metadata/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/widgets/metadata `
+    https://github.com/Zuoqiu-Yingyi/widget-metadata.git `
+    dev