+++ scripts/git-subtree/metadata/split.ps1
@@ -1,0 +1,3 @@
+git subtree split `
+    -P workspace/widgets/metadata `
+    --rejoin